rename record_config to recorder_config
